backend-python/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==2.0.2
pandas==2.1.3
numpy==1.25.2
matplotlib==3.8.2
seaborn==0.13.0
ta==0.10.2
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
schedule==1.2.0
ShoonyaApi-py==0.1.28
yfinance==0.2.28
